+++ R/salesforcer.R
@@ -14,4 +14,8 @@
 NULL
 
 ## quiets concerns of R CMD check re: the .'s that appear in pipelines
-if(getRversion() >= "2.15.1")  utils::globalVariables(c(".", "id", "success"))
+if(getRversion() >= "2.15.1")  utils::globalVariables(c(".", 
+                                                        "id", 
+                                                        "success", 
+                                                        "attributes.type", 
+                                                        "sf:type"))