+++ tests/testthat/test-compatibility.R
@@ -146,13 +146,31 @@ test_that("testing rforcecom.getServerTimestamp compatibility", {
                round(result2, units = "mins"))
 })
 
-# test_that("testing rforcecom.retrieve compatibility", {
-#   
-# })
+test_that("testing rforcecom.retrieve compatibility", {
+  
+  objectName <- "Account"
+  fields <- c("name", "Industry", "AnnualRevenue")
+  
+  result1 <- RForcecom::rforcecom.retrieve(session, objectName, fields, limit = 5)
+  result2 <- salesforcer::rforcecom.retrieve(session, objectName, fields, limit = 5)
+  
+  expect_equal(names(result1), c("Name", "Industry", "AnnualRevenue"))
+  expect_is(result1, "data.frame")
+  expect_is(result2, "data.frame")
+  expect_equal(sort(names(result1)), sort(names(result2)))
+  expect_equal(nrow(result1), nrow(result2))
+})
 
-# test_that("testing rforcecom.search compatibility", {
-#   
-# })
+test_that("testing rforcecom.search compatibility", {
+  
+  search_string <- "(336)"
+  result1 <- RForcecom::rforcecom.search(session, search_string)
+  result2 <- salesforcer::rforcecom.search(session, search_string)
+
+  expect_null(result1)
+  # rforcecom.search has a bug that wont return right data
+  expect_is(result2, "data.frame")
+})
 
 # test_that("testing rforcecom.bulkAction compatibility", {
 #